requirements.txt
asgiref==3.9.1
certifi==2025.8.3
Django==5.2.6
django-filter==25.1
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
drf-yasg==1.21.10
elastic-transport==9.1.0
elasticsearch==9.1.1
inflection==0.5.1
Markdown==3.9
packaging==25.0
pillow==11.3.0
psycopg2-binary==2.9.10
PyJWT==2.10.1
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
pytz==2025.2
PyYAML==6.0.2
six==1.17.0
sqlparse==0.5.3
typing_extensions==4.15.0
uritemplate==4.2.0
urllib3==2.5.0
